lib: fix efi_time_to_ctime overflow
When computing the ctime based on the intermediate (days, Hour, ...)
values, they need to be casted to UINT64 otherwise they overflow.

It was only an issue for time later of equal to year 2106.

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-71456
Reviewed-on: https://android.intel.com:443/651072

+++ libkernelflinger/lib.c
@@ -1016,8 +1016,10 @@ UINT64 efi_time_to_ctime(EFI_TIME *time)
         for (i = 0; i + 1 < time->Month; i++)
                 days += DAY_OF_MONTH[i];
 
-        return (days * 24 * 3600) + (time->Hour * 3600)
-                + (time->Minute * 60) + time->Second;
+        return ((UINT64)days * 24 * 3600) +
+                ((UINT64)time->Hour * 3600) +
+                ((UINT64)time->Minute * 60) +
+                (UINT64)time->Second;
 }
 
 VOID cpuid(UINT32 op, UINT32 reg[4])
